pyproject.toml
[project]
name = "autohome"
version = "0.1.0"
description = "Real estate metrics automation: Idealista scraping → condition engine → Clientify actions"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.49",
    "httpx>=0.27",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "rich>=13.9",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-playwright>=0.6",
    "ruff>=0.8",
    "pre-commit>=4.0",
]

[project.scripts]
autohome = "autohome.cli:main"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S101"]  # Allow assert in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
